# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZuSu Trading API",
    "version": "0.1.0",
    "description": "Day-trading automation platform. Phase 1 covers identity, portfolios, the demo broker, the trading gate and the audit log. Endpoints that would place orders do not exist yet — they arrive with the risk engine and order manager."
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zusu_at"
      }
    },
...
  },
  "paths": {
...
    "/api/broker/{id}/account": {
      "get": {
        "summary": "The broker’s own view of the account behind a portfolio",
        "tags": ["broker"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accountId": {
                      "type": "string"
                    },
                    "environment": {
                      "type": "string",
                      "enum": ["DEMO", "PAPER", "LIVE"]
                    },
                    "currency": {
                      "type": "string"
                    },
                    "cash": {
                      "type": "string",
                      "pattern": "^-?\\d+(\\.\\d+)?$"
                    },
                    "buyingPower": {
                      "type": "string",
                      "pattern": "^-?\\d+(\\.\\d+)?$"
                    },
                    "equity": {
                      "type": "string",
                      "pattern": "^-?\\d+(\\.\\d+)?$"
                    },
                    "isPatternDayTrader": {
                      "type": "boolean"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "accountId",
                    "environment",
                    "currency",
                    "cash",
                    "buyingPower",
                    "equity",
                    "isPatternDayTrader",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/broker/{id}/positions": {
      "get": {
        "summary": "Positions as the broker reports them (the reconciliation counterpart)",
        "tags": ["broker"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "symbol": {
                        "type": "string"
                      },
                      "assetClass": {
                        "type": "string"
                      },
                      "quantity": {
                        "type": "string",
                        "pattern": "^-?\\d+(\\.\\d+)?$"
                      },
                      "averageEntryPrice": {
                        "type": "string",
                        "pattern": "^-?\\d+(\\.\\d+)?$"
                      },
                      "markPrice": {
                        "type": "string",
                        "pattern": "^-?\\d+(\\.\\d+)?$"
                      },
                      "marketValue": {
                        "type": "string",
                        "pattern": "^-?\\d+(\\.\\d+)?$"
                      },
                      "unrealizedPnl": {
                        "type": "string",
                        "pattern": "^-?\\d+(\\.\\d+)?$"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    },
                    "required": [
                      "symbol",
                      "assetClass",
                      "quantity",
                      "averageEntryPrice",
                      "markPrice",
                      "marketValue",
                      "unrealizedPnl",
                      "updatedAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/broker/{id}/quote/{symbol}": {
      "get": {
        "summary": "A quote from the portfolio environment’s market-data source",
        "tags": ["broker"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 24,
              "pattern": "^[A-Z0-9.\\-/:]+$"
            },
            "in": "path",
            "name": "symbol",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "symbol": {
                      "type": "string"
                    },
                    "provider": {
                      "type": "string"
                    },
                    "price": {
                      "type": "string",
                      "pattern": "^-?\\d+(\\.\\d+)?$"
                    },
                    "bid": {
                      "type": "string",
                      "pattern": "^-?\\d+(\\.\\d+)?$"
                    },
                    "ask": {
                      "type": "string",
                      "pattern": "^-?\\d+(\\.\\d+)?$"
                    },
                    "volume": {
                      "type": "string",
                      "pattern": "^-?\\d+(\\.\\d+)?$"
                    },
                    "marketSession": {
                      "type": "string"
                    },
                    "sourceTimestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "receivedTimestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "symbol",
                    "provider",
                    "price",
                    "bid",
                    "ask",
                    "volume",
                    "marketSession",
                    "sourceTimestamp",
                    "receivedTimestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "this deployment"
    }
  ],
...
}